requirements.txt
pydantic
pyyaml
pytest
fastapi
uvicorn